Use Path.DirectorySeparatorChar for portability
Replaced usages of literal backslash in paths with
Path.DirectorySeparatorChar in order to allow paths to be constructed
properly when not running on a Windows system.

Before this change when runing using Mono on a Mac, output files would
have backslashes in the file name instead of being placed in the
apropriate directory structure. After this change, files are placed
properly.